HCA move env files under hca/envs.  Move tests to tests/
